Most recent version of ExponentialWavefunctionStepper.* files that include a merged exponential and code simplification, i.e. I eliminated some redundancies between the ETRS and AETRS versions of the code. This version has been tested.
